Add task property annotations
Due to the structure of the code and not using the Provider API it is
not possible to supply default values for all task input properties.
If no value for such properties is specified and the generated default
value changes, this is not detected and the task is not marked as out-of-date.
In such a case, a manual clean and full rebuild is required.

The following tasks and properties are affected:

- vaadinCompile
  - outputDirectory
  - widgetset
  - widgetsetGenerator
- vaadinThemeCompile
  - themesDirectory